+++ mod-source-record-manager-server/src/main/java/org/folio/services/migration/impl/AuthorityMappingChronSubdivisionCustomMigration.java
@@ -1,0 +1,40 @@
+package org.folio.services.migration.impl;
+
+import static org.folio.Record.RecordType.MARC_AUTHORITY;
+
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
+import org.folio.services.MappingRuleService;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthorityMappingChronSubdivisionCustomMigration extends BaseMappingRulesMigration {
+
+  private static final JsonArray SUBFIELDS = JsonArray.of("v", "x", "y", "z");
+  private static final String TARGET_182 = "chronSubdivision";
+  private static final String TARGET_482 = "sftChronSubdivision";
+  private static final String TARGET_582 = "saftChronSubdivision";
+  private static final String DESCRIPTION_182 = "Heading chron subdivision";
+  private static final String DESCRIPTION_482 = "See from tracing chron subdivision";
+  private static final String DESCRIPTION_582 = "See also from tracing chron subdivision";
+  private static final String TAG_182 = "182";
+  private static final String TAG_482 = "482";
+  private static final String TAG_582 = "582";
+  private static final String FEATURE_VERSION = "3.10.0";
+  private static final String DESCRIPTION = "Authority mapping rules: add rules for chron subdivision fields";
+
+  protected AuthorityMappingChronSubdivisionCustomMigration(MappingRuleService mappingRuleService) {
+    super(MARC_AUTHORITY, FEATURE_VERSION, DESCRIPTION, mappingRuleService);
+  }
+
+  @Override
+  protected String updateRules(JsonObject rules) {
+    var field182 = createField(TARGET_182, DESCRIPTION_182, SUBFIELDS, EMPTY_RULES);
+    var field482 = createField(TARGET_482, DESCRIPTION_482, SUBFIELDS, EMPTY_RULES);
+    var field582 = createField(TARGET_582, DESCRIPTION_582, SUBFIELDS, EMPTY_RULES);
+    addFieldIfNotExists(rules, TAG_182, field182);
+    addFieldIfNotExists(rules, TAG_482, field482);
+    addFieldIfNotExists(rules, TAG_582, field582);
+    return sortRules(rules).encode();
+  }
+}